+++ crates/re_viewer/src/lib.rs
@@ -10,10 +10,8 @@ mod loading;
 #[cfg(not(target_arch = "wasm32"))]
 mod profiler;
 mod remote_viewer_app;
-mod rerun_menu;
 mod screenshotter;
 mod store_hub;
-mod top_panel;
 mod ui;
 mod viewer_analytics;
