+++ crates/re_viewer/src/ui/mod.rs
@@ -1,11 +1,15 @@
 mod blueprint;
 mod blueprint_load;
 mod blueprint_sync;
+mod rerun_menu;
 mod selection_history_ui;
+mod top_panel;
 
 pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
 
 // ----
 
-pub(crate) use self::blueprint::Blueprint;
+pub(crate) use {
+    self::blueprint::Blueprint, self::rerun_menu::rerun_menu_button_ui, self::top_panel::top_panel,
+};
