+++ crates/re_viewer/src/ui/rerun_menu.rs
